backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
gunicorn
dj-database-url>=2.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
